Fix optimize_image (fix #87)
Fix a miscalculaton of end of partition parameter for parted. It shall be the
last byte of the partition, not start byte + size. This made parted fail when
the partition wasn't shrunk by resize2fs.

Avoid calling parted if the partition isn't shrunk by resize2fs.

Fix a bug where `cleanup_image.sh` script was not aborting in case of errors,
thus not revealing a corruption of generated image as reported in #87.

Fix a bug where cleanup_image.sh failed using images with more than 10
partitions because of a lazy glob.

Signed-off-by: Paul Guyot <pguyot@kallisys.net>